requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Smorest==0.44.0
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.1
Flask-Limiter==3.6.0
python-dotenv==1.0.1
bcrypt==4.1.3
gunicorn==22.0.0
psycopg2-binary # Cần cho PostgreSQL ở production